CPU-separation: extract device-specific code from `logging.cpp`
This includes moving Tap and KeyValue loggers as cpu-only.
Inheritance order is changed, and now both Tap and KeyValue inherit
from Yaml.
